bring in the latest juju/txn.
This brings in the Oracles to provide in-memory or on-disk tracking of
transactions to be pruned.
It also brings in the 'cleanup' logic that removes transactions that have
been completed from the txn-queues rather than not touching them.
And brings in improvements to remove documents from txns.stash that
only refer to transactions that have been marked completed.
Since we now purge all completed transactions from all txn-queues,
we no longer need a specific stage for machines.