Trim transaction queues that are too long.
Handle the specific case of documents that are touched very frequently.
The most recent cases involved transactions that touched a single document,
and caused their txn queue to get out of hand. This fix just makes it
so mgopurge handles very narrow transactions (affecting only the doc
in question), and removes those transactions and the tokens from the
document.