Support handling txns with long queues that involve inserts.
It turned out that you can get too-long txn-queues on a transaction that
might be inserting a document, which meant the referenced document was
still in the stash, not in the actual table it was supposed to end up
in.
And that meant we couldn't confirm that all docs were involved in the
long txn-queue so we didn't know if it was safe to purge the queue.

This update changes it so we check in txns.stash and give it special
handling since the docs there are keyed differently.