Also cleanup the Makefile.
This allows us to build using 'snapcraft' again. It isn't perfect, but
it is a step in the right direction.

Note that mgpurge can only be built using go 1.12 right now (1.13
complains about a missing main function.)
Ideally we would get off of GB entirely, and just use go.mod.